add new data

+++ database/migrations/2021_09_30_101349_create_discounts_table.php
@@ -15,7 +15,7 @@ class CreateDiscountsTable extends Migration
     {
         Schema::create('discounts', function (Blueprint $table) {
             $table->id();
-            $table->integer('discount');
+            $table->bigInteger('discount');
             $table->string('type');
             $table->foreignId('group_user_id')->nullable()->constrained('roles')->onUpdate('cascade')->onDelete('cascade');
             $table->foreignId('user_id')->nullable()->constrained('users')->onUpdate('cascade')->onDelete('cascade');
